Losing lines, wrong count for splitted parts
It seems there was a serious bug on splitting the original file,
as it was loosing lines, so they were never making it into the splitted
parts.

Also the splitted parts count was wrong in cases when there was still
something in the buffer after the original file was fully read.